allowed sweeping reduced range
Added #defines to allow sweeping a reduced range, because some servos (such as mine) are unhappy at the extremities. My servo makes growling noises and vibrates more strongly if I tell it to go to less than 30 degrees or more than 160 degrees, so I wanted to modify the program to allow that. 

At first I reduced the size of the array for storing values but that was a mistake and caused the program to behave strangely, so I added extra print()s while trying to figure out why. Finally I realized that the index in the array of measurements was absolute, so could easily be outside of the reduced range. 

Accessing an array outside of the size of the array causes all sorts of strange behavior.